handle offline mints that are unknown
